+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import flutter_tor
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  FlutterTorPlugin.register(with: registry.registrar(forPlugin: "FlutterTorPlugin"))
+}